App Routing - Migration Video Production pages
Affected Route: `/consulting/video-production/*`

Fixed #3162

Screenshot:
TODO
