+++ components/blocks/carousel.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import * as React from "react";
