+++ components/usergroup/organizer.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Image from "next/image";
 import { useState } from "react";
 import type { Template } from "tinacms";
